Final
final touches, tutorial

+++ login/load_profile.php
@@ -10,7 +10,7 @@
 	$result = mysqli_query($con, $SQL); //query result
 	$userName = $uname; //username retreived from session variable
 	$e = mysqli_fetch_assoc($result)['userEmail']; //pull email from DB
-	$no = mysqli_fetch_assoc($result)['user_id'];//pull ID
+	$no = (int)mysqli_fetch_assoc($result)['user_id'];//pull ID
 	$userEmail = $e;
 	$userId = $no;
